+++ src/static/dolphin/os/OSStopwatch.c
@@ -18,7 +18,7 @@ void OSStartStopwatch(struct OSStopwatch* sw)
 
 void OSStopStopwatch(struct OSStopwatch* sw)
 {
-	long long interval;
+	s64 interval;
 
 	if (sw->running != 0) {
 		interval = OSGetTime() - sw->last;
@@ -34,9 +34,9 @@ void OSStopStopwatch(struct OSStopwatch* sw)
 	}
 }
 
-long long OSCheckStopwatch(struct OSStopwatch* sw)
+s64 OSCheckStopwatch(struct OSStopwatch* sw)
 {
-	long long currTotal;
+	s64 currTotal;
 
 	currTotal = sw->total;
 	if (sw->running != 0) {
